+++ stats.cpp
@@ -1,7 +1,7 @@
 #include "stats.h"
 
-int calculePourcentage(int valeur , int total) {   
-   int pourcentage ;
+float calculePourcentage(int valeur , int total) {   
+   float pourcentage ;
    //C++ a besoin de type float pour avoir une division précise
    //Nous retournons quand même un résultat en int , selon le cahier des charges
    pourcentage = ((float) valeur / (float)total)  * 100;
@@ -9,7 +9,7 @@ int calculePourcentage(int valeur , int total) {
 }
 
 int calculeTotal(int valeur, int total){
-   int reponse;
+   float reponse;
    reponse = (valeur / total);
    return reponse;  
 }
