+++ stats.h
@@ -1,7 +1,7 @@
 #ifndef STATS_H
 #define STATS_H
 
-int calculePourcentage( int valeur , int total );
+float calculePourcentage( int valeur , int total );
 
 int calculeTotal(int valeur, int total);
